+++ examples/text_layouter_shapes.rb
@@ -1,8 +1,9 @@
-# ## Text Box Shapes
+# ## Text Layouter - Shapes
 #
-# The [HexaPDF::Layout::TextBox] class can be used to easily lay out text, not
-# limiting the area to a rectangle but any shape. There is only one restriction:
-# In the case of arbitrary shapes the vertical alignment has to be "top".
+# The [HexaPDF::Layout::TextLayouter] class can be used to easily lay out text,
+# not limiting the area to a rectangle but any shape. There is only one
+# restriction: In the case of arbitrary shapes the vertical alignment has to be
+# "top".
 #
 # Arbitrary shapes boil down to varying line widths and horizontal offsets from
 # left. Imagine a circle: If text is fit in a circle, the line widths start at
@@ -12,18 +13,20 @@
 #
 # Both, the line widths and the horizontal offsets can be calculated given a
 # certain height, and this is exactly what HexaPDF uses. If the `width` argument
-# to [HexaPDF::Layout::TextBox::new] is an object responding to #call (e.g. a
-# lambda), it is used for determining the line widths. And the `x_offsets`
+# to [HexaPDF::Layout::TextLayouter::new] is an object responding to #call (e.g.
+# a lambda), it is used for determining the line widths. And the `x_offsets`
 # argument can be used in a similar way for the horizontal offsets.
 #
 # This example shows text layed out in various shapes, using the above mentioned
 # techniques.
 #
 # Usage:
-# : `ruby text_box_shapes.rb`
+# : `ruby text_layouter_shapes.rb`
 #
 
 require 'hexapdf'
+
+include HexaPDF::Layout
 
 doc = HexaPDF::Document.new
 page = doc.pages.add
@@ -58,31 +61,31 @@ left_half_circle_offsets = lambda do |height, line_height|
 end
 
 # Left: right half circle
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: half_circle_widths,
-                                      height: radius * 2,
-                                      font: doc.fonts.load("Times"))
-box.draw(canvas, 0, circle_top)
+layouter = TextLayouter.create(sample_text,
+                               width: half_circle_widths,
+                               height: radius * 2,
+                               font: doc.fonts.load("Times"))
+layouter.draw(canvas, 0, circle_top)
 canvas.circle(0, circle_top - radius, radius).stroke
 
 # Center: full circle
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: circle_widths,
-                                      x_offsets: left_half_circle_offsets,
-                                      height: radius * 2,
-                                      font: doc.fonts.load("Times"),
-                                      align: :justify)
-box.draw(canvas, page.box(:media).width / 2.0 - radius, circle_top)
+layouter = TextLayouter.create(sample_text,
+                               width: circle_widths,
+                               x_offsets: left_half_circle_offsets,
+                               height: radius * 2,
+                               font: doc.fonts.load("Times"),
+                               align: :justify)
+layouter.draw(canvas, page.box(:media).width / 2.0 - radius, circle_top)
 canvas.circle(page.box(:media).width / 2.0, circle_top - radius, radius).stroke
 
 # Right: left half circle
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: half_circle_widths,
-                                      x_offsets: left_half_circle_offsets,
-                                      height: radius * 2,
-                                      font: doc.fonts.load("Times"),
-                                      align: :right)
-box.draw(canvas, page.box(:media).width - radius, circle_top)
+layouter = TextLayouter.create(sample_text,
+                               width: half_circle_widths,
+                               x_offsets: left_half_circle_offsets,
+                               height: radius * 2,
+                               font: doc.fonts.load("Times"),
+                               align: :right)
+layouter.draw(canvas, page.box(:media).width - radius, circle_top)
 canvas.circle(page.box(:media).width, circle_top - radius, radius).stroke
 
 
@@ -107,37 +110,37 @@ left_half_diamond_offsets = lambda do |height, line_height|
 end
 
 # Left: right half diamond
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: half_diamond_widths,
-                                      height: 2 * diamond_width,
-                                      font: doc.fonts.load("Times"))
-box.draw(canvas, 0, diamond_top)
+layouter = TextLayouter.create(sample_text,
+                               width: half_diamond_widths,
+                               height: 2 * diamond_width,
+                               font: doc.fonts.load("Times"))
+layouter.draw(canvas, 0, diamond_top)
 canvas.polyline(0, diamond_top, diamond_width, diamond_top - diamond_width,
-               0, diamond_top - 2 * diamond_width).stroke
+                0, diamond_top - 2 * diamond_width).stroke
 
 # Center: full diamond
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: full_diamond_widths,
-                                      x_offsets: left_half_diamond_offsets,
-                                      height: 2 * diamond_width,
-                                      font: doc.fonts.load("Times"),
-                                      align: :justify)
+layouter = TextLayouter.create(sample_text,
+                               width: full_diamond_widths,
+                               x_offsets: left_half_diamond_offsets,
+                               height: 2 * diamond_width,
+                               font: doc.fonts.load("Times"),
+                               align: :justify)
 left = page.box(:media).width / 2.0 - diamond_width
-box.draw(canvas, left, diamond_top)
+layouter.draw(canvas, left, diamond_top)
 canvas.polyline(left + diamond_width, diamond_top,
                 left + 2 * diamond_width, diamond_top - diamond_width,
                 left + diamond_width, diamond_top - 2 * diamond_width,
                 left, diamond_top - diamond_width).close_subpath.stroke
 
 # Right: left half diamond
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: half_diamond_widths,
-                                      x_offsets: left_half_diamond_offsets,
-                                      height: 2 * diamond_width,
-                                      font: doc.fonts.load("Times"),
-                                      align: :right)
+layouter = TextLayouter.create(sample_text,
+                               width: half_diamond_widths,
+                               x_offsets: left_half_diamond_offsets,
+                               height: 2 * diamond_width,
+                               font: doc.fonts.load("Times"),
+                               align: :right)
 middle = page.box(:media).width
-box.draw(canvas, middle - diamond_width, diamond_top)
+layouter.draw(canvas, middle - diamond_width, diamond_top)
 canvas.polyline(middle, diamond_top,
                 middle - diamond_width, diamond_top - diamond_width,
                 middle, diamond_top - 2 * diamond_width).stroke
@@ -154,13 +157,13 @@ end
 sine_wave_widths = lambda do |height, line_height|
   sine_wave_height + 100 + sine_wave_offsets.call(height, line_height) * -2
 end
-box = HexaPDF::Layout::TextBox.create(sample_text,
-                                      width: sine_wave_widths,
-                                      x_offsets: sine_wave_offsets,
-                                      height: sine_wave_height,
-                                      font: doc.fonts.load("Times"),
-                                      align: :justify)
+layouter = TextLayouter.create(sample_text,
+                               width: sine_wave_widths,
+                               x_offsets: sine_wave_offsets,
+                               height: sine_wave_height,
+                               font: doc.fonts.load("Times"),
+                               align: :justify)
 middle = page.box(:media).width / 2.0
-box.draw(canvas, middle - (sine_wave_height + 100) / 2, sine_wave_top)
+layouter.draw(canvas, middle - (sine_wave_height + 100) / 2, sine_wave_top)
 
-doc.write("text_box_shapes.pdf", optimize: true)
+doc.write("text_layouter_shapes.pdf", optimize: true)