+++ test/hexapdf/layout/test_text_layouter.rb
@@ -5,28 +5,28 @@ require 'hexapdf/layout'
 require 'hexapdf/document'
 require_relative "../content/common"
 
-module TestTextBoxHelpers
+module TestTextLayouterHelpers
   def boxes(*dims)
     dims.map do |width, height|
       box = HexaPDF::Layout::InlineBox.new(width, height || 0) {}
-      HexaPDF::Layout::TextBox::Box.new(box)
+      HexaPDF::Layout::TextLayouter::Box.new(box)
     end
   end
 
   def glue(width)
-    HexaPDF::Layout::TextBox::Glue.new(HexaPDF::Layout::InlineBox.new(width, 0) {})
+    HexaPDF::Layout::TextLayouter::Glue.new(HexaPDF::Layout::InlineBox.new(width, 0) {})
   end
 
   def penalty(penalty, item = nil)
     if item
-      HexaPDF::Layout::TextBox::Penalty.new(penalty, item.width, item: item)
+      HexaPDF::Layout::TextLayouter::Penalty.new(penalty, item.width, item: item)
     else
-      HexaPDF::Layout::TextBox::Penalty.new(penalty)
+      HexaPDF::Layout::TextLayouter::Penalty.new(penalty)
     end
   end
 
   def assert_box(obj, item)
-    assert_kind_of(HexaPDF::Layout::TextBox::Box, obj)
+    assert_kind_of(HexaPDF::Layout::TextLayouter::Box, obj)
     if obj.item.kind_of?(HexaPDF::Layout::InlineBox)
       assert_same(item, obj.item)
     else
@@ -36,12 +36,12 @@ module TestTextBoxHelpers
   end
 
   def assert_glue(obj, fragment)
-    assert_kind_of(HexaPDF::Layout::TextBox::Glue, obj)
+    assert_kind_of(HexaPDF::Layout::TextLayouter::Glue, obj)
     assert_same(fragment.style, obj.item.style)
   end
 
   def assert_penalty(obj, penalty, item = nil)
-    assert_kind_of(HexaPDF::Layout::TextBox::Penalty, obj)
+    assert_kind_of(HexaPDF::Layout::TextLayouter::Penalty, obj)
     assert_equal(penalty, obj.penalty)
     if item
       assert_same(item.style, obj.item.style)
@@ -50,13 +50,13 @@ module TestTextBoxHelpers
   end
 end
 
-describe HexaPDF::Layout::TextBox::SimpleTextSegmentation do
-  include TestTextBoxHelpers
+describe HexaPDF::Layout::TextLayouter::SimpleTextSegmentation do
+  include TestTextLayouterHelpers
 
   before do
     @doc = HexaPDF::Document.new
     @font = @doc.fonts.load("Times")
-    @obj = HexaPDF::Layout::TextBox::SimpleTextSegmentation
+    @obj = HexaPDF::Layout::TextLayouter::SimpleTextSegmentation
   end
 
   def setup_fragment(text, style = nil)
@@ -109,10 +109,10 @@ describe HexaPDF::Layout::TextBox::SimpleTextSegmentation do
     assert_equal(17, result.size)
     [1, 3, 5, 7, 9, 11, 13].each do |index|
       assert_penalty(result[index],
-                     HexaPDF::Layout::TextBox::Penalty::MandatoryParagraphBreak.penalty)
+                     HexaPDF::Layout::TextLayouter::Penalty::MandatoryParagraphBreak.penalty)
     end
     assert_penalty(result[15],
-                   HexaPDF::Layout::TextBox::Penalty::MandatoryLineBreak.penalty)
+                   HexaPDF::Layout::TextLayouter::Penalty::MandatoryLineBreak.penalty)
   end
 
   it "insert a standard penalty after a hyphen" do
@@ -121,7 +121,7 @@ describe HexaPDF::Layout::TextBox::SimpleTextSegmentation do
     result = @obj.call([frag])
     assert_equal(12, result.size)
     [1, 3, 5, 9].each do |index|
-      assert_penalty(result[index], HexaPDF::Layout::TextBox::Penalty::Standard.penalty)
+      assert_penalty(result[index], HexaPDF::Layout::TextLayouter::Penalty::Standard.penalty)
     end
   end
 
@@ -140,7 +140,7 @@ describe HexaPDF::Layout::TextBox::SimpleTextSegmentation do
 
     result = @obj.call([frag])
     assert_equal(3, result.size)
-    assert_penalty(result[1], HexaPDF::Layout::TextBox::Penalty::Standard.penalty, hyphen)
+    assert_penalty(result[1], HexaPDF::Layout::TextLayouter::Penalty::Standard.penalty, hyphen)
   end
 
   it "insert a prohibited break penalty for non-breaking spaces" do
@@ -149,7 +149,7 @@ describe HexaPDF::Layout::TextBox::SimpleTextSegmentation do
 
     result = @obj.call([frag])
     assert_equal(3, result.size)
-    assert_penalty(result[1], HexaPDF::Layout::TextBox::Penalty::ProhibitedBreak.penalty, space)
+    assert_penalty(result[1], HexaPDF::Layout::TextLayouter::Penalty::ProhibitedBreak.penalty, space)
   end
 end
 
@@ -159,7 +159,7 @@ end
 module CommonLineWrappingTests
   extend Minitest::Spec::DSL
 
-  include TestTextBoxHelpers
+  include TestTextLayouterHelpers
 
   it "breaks before a box if it doesn't fit onto the line anymore" do
     rest, lines = call(boxes(25, 50, 25, 10))
@@ -287,9 +287,9 @@ module CommonLineWrappingTests
 
 end
 
-describe HexaPDF::Layout::TextBox::SimpleLineWrapping do
+describe HexaPDF::Layout::TextLayouter::SimpleLineWrapping do
   before do
-    @obj = HexaPDF::Layout::TextBox::SimpleLineWrapping
+    @obj = HexaPDF::Layout::TextLayouter::SimpleLineWrapping
   end
 
   describe "fixed width wrapping" do
@@ -361,8 +361,8 @@ describe HexaPDF::Layout::TextBox::SimpleLineWrapping do
   end
 end
 
-describe HexaPDF::Layout::TextBox do
-  include TestTextBoxHelpers
+describe HexaPDF::Layout::TextLayouter do
+  include TestTextLayouterHelpers
 
   before do
     @doc = HexaPDF::Document.new
@@ -371,60 +371,61 @@ describe HexaPDF::Layout::TextBox do
   end
 
   it "creates an instance from text and options" do
-    box = HexaPDF::Layout::TextBox.create("T", font: @font, width: 100, height: 100)
-    assert_equal(1, box.items.length)
-    assert_equal(@font.decode_utf8("T"), box.items[0].item.items)
+    layouter = HexaPDF::Layout::TextLayouter.create("T", font: @font, width: 100, height: 100)
+    assert_equal(1, layouter.items.length)
+    assert_equal(@font.decode_utf8("T"), layouter.items[0].item.items)
   end
 
   it "doesn't run the text segmentation algorithm on already segmented items" do
     item = HexaPDF::Layout::InlineBox.new(20, 0) {}
-    box = HexaPDF::Layout::TextBox.new(items: [item], width: 100, height: 100)
-    items = box.items
+    layouter = HexaPDF::Layout::TextLayouter.new(items: [item], width: 100, height: 100)
+    items = layouter.items
     assert_equal(1, items.length)
     assert_box(items[0], item)
 
-    box.items = items
-    assert_same(items, box.items)
+    layouter.items = items
+    assert_same(items, layouter.items)
   end
 
   describe "fit" do
     it "handles text indentation" do
       items = boxes([20, 20], [20, 20], [20, 20]) +
-        [HexaPDF::Layout::TextBox::Penalty::MandatoryParagraphBreak] +
+        [HexaPDF::Layout::TextLayouter::Penalty::MandatoryParagraphBreak] +
         boxes([40, 20]) + [glue(20)] +
-        boxes(*([[20, 20]] * 4)) + [HexaPDF::Layout::TextBox::Penalty::MandatoryLineBreak] +
+        boxes(*([[20, 20]] * 4)) + [HexaPDF::Layout::TextLayouter::Penalty::MandatoryLineBreak] +
         boxes(*([[20, 20]] * 4))
       @style.text_indent = 20
 
       [60, proc { 60 }].each do |width|
-        box = HexaPDF::Layout::TextBox.new(items: items, width: width, style: @style)
-        rest, = box.fit
-        assert_equal([40, 20, 40, 60, 20, 60, 20], box.lines.map(&:width))
-        assert_equal([20, 0, 20, 0, 0, 0, 0], box.lines.map(&:x_offset))
+        layouter = HexaPDF::Layout::TextLayouter.new(items: items, width: width, style: @style)
+        rest, = layouter.fit
+        assert_equal([40, 20, 40, 60, 20, 60, 20], layouter.lines.map(&:width))
+        assert_equal([20, 0, 20, 0, 0, 0, 0], layouter.lines.map(&:x_offset))
         assert(rest.empty?)
       end
     end
 
     it "fits using unlimited height" do
-      box = HexaPDF::Layout::TextBox.new(items: boxes(*([[20, 20]] * 100)), width: 20,
-                                         style: @style)
-      rest, height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: boxes(*([[20, 20]] * 100)), width: 20,
+                                                   style: @style)
+      rest, height = layouter.fit
       assert(rest.empty?)
       assert_equal(20 * 100, height)
     end
 
     it "fits using a limited height" do
-      box = HexaPDF::Layout::TextBox.new(items: boxes(*([[20, 20]] * 100)), width: 20, height: 100,
-                                         style: @style)
-      rest, height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: boxes(*([[20, 20]] * 100)), width: 20,
+                                                   height: 100, style: @style)
+      rest, height = layouter.fit
       assert_equal(95, rest.count)
       assert_equal(100, height)
     end
 
     it "takes line spacing into account when calculating the height" do
-      box = HexaPDF::Layout::TextBox.new(items: boxes(*([[20, 20]] * 5)), width: 20, style: @style)
-      box.style.line_spacing = :double
-      rest, height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: boxes(*([[20, 20]] * 5)), width: 20,
+                                                   style: @style)
+      layouter.style.line_spacing = :double
+      rest, height = layouter.fit
       assert(rest.empty?)
       assert_equal(20 * (5 + 4), height)
     end
@@ -432,26 +433,26 @@ describe HexaPDF::Layout::TextBox do
     it "handles empty lines" do
       items = boxes([20, 20]) + [penalty(-5000)] + boxes([30, 20]) + [penalty(-5000)] * 2 +
         boxes([20, 20]) + [penalty(-5000)] * 2
-      box = HexaPDF::Layout::TextBox.new(items: items, width: 30, style: @style)
-      rest, height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: items, width: 30, style: @style)
+      rest, height = layouter.fit
       assert(rest.empty?)
-      assert_equal(5, box.lines.count)
+      assert_equal(5, layouter.lines.count)
       assert_equal(20 + 20 + 9 + 20 + 9, height)
     end
 
     describe "fixed width" do
       it "stops if an item is wider than the available width, with unlimited height" do
-        box = HexaPDF::Layout::TextBox.new(items: boxes([20, 20], [50, 20]), width: 30,
-                                           style: @style)
-        rest, height = box.fit
+        layouter = HexaPDF::Layout::TextLayouter.new(items: boxes([20, 20], [50, 20]), width: 30,
+                                                     style: @style)
+        rest, height = layouter.fit
         assert_equal(1, rest.count)
         assert_equal(20, height)
       end
 
       it "stops if an item is wider than the available width, with limited height" do
-        box = HexaPDF::Layout::TextBox.new(items: boxes([20, 20], [50, 20]), width: 30, height: 100,
-                                           style: @style)
-        rest, height = box.fit
+        layouter = HexaPDF::Layout::TextLayouter.new(items: boxes([20, 20], [50, 20]), width: 30,
+                                                     height: 100, style: @style)
+        rest, height = layouter.fit
         assert_equal(1, rest.count)
         assert_equal(20, height)
       end
@@ -465,12 +466,12 @@ describe HexaPDF::Layout::TextBox do
           else 40
           end
         end
-        box = HexaPDF::Layout::TextBox.new(items: boxes([20, 18]), width: width_block,
-                                           height: 100, style: @style)
-        rest, height = box.fit
+        layouter = HexaPDF::Layout::TextLayouter.new(items: boxes([20, 18]), width: width_block,
+                                                     height: 100, style: @style)
+        rest, height = layouter.fit
         assert(rest.empty?)
-        assert_equal(1, box.lines.count)
-        assert_equal(24, box.lines[0].y_offset)
+        assert_equal(1, layouter.lines.count)
+        assert_equal(24, layouter.lines[0].y_offset)
         assert_equal(42, height)
       end
 
@@ -482,14 +483,14 @@ describe HexaPDF::Layout::TextBox do
             40
           end
         end
-        box = HexaPDF::Layout::TextBox.new(items: boxes(*([[20, 18]] * 7)), width: width_block,
-                                           height: 100, style: @style)
-        rest, height = box.fit
+        layouter = HexaPDF::Layout::TextLayouter.new(items: boxes(*([[20, 18]] * 7)),
+                                                     width: width_block, height: 100, style: @style)
+        rest, height = layouter.fit
         assert_equal(1, rest.count)
-        assert_equal(3, box.lines.count)
-        assert_equal(0, box.lines[0].y_offset)
-        assert_equal(18, box.lines[1].y_offset)
-        assert_equal(48, box.lines[2].y_offset)
+        assert_equal(3, layouter.lines.count)
+        assert_equal(0, layouter.lines[0].y_offset)
+        assert_equal(18, layouter.lines[1].y_offset)
+        assert_equal(48, layouter.lines[2].y_offset)
         assert_equal(84, height)
       end
     end
@@ -497,37 +498,37 @@ describe HexaPDF::Layout::TextBox do
     it "post-processes lines for justification if needed" do
       frag10 = HexaPDF::Layout::TextFragment.create(" ", font: @font)
       frag10.items.freeze
-      frag10b = HexaPDF::Layout::TextBox::Box.new(frag10)
+      frag10b = HexaPDF::Layout::TextLayouter::Box.new(frag10)
       frag20 = HexaPDF::Layout::TextFragment.create(" ", font: @font, font_size: 20)
-      frag20b = HexaPDF::Layout::TextBox::Box.new(frag20)
+      frag20b = HexaPDF::Layout::TextLayouter::Box.new(frag20)
       items = boxes(20, 20, 20, 20, 30).insert(1, frag10b).insert(3, frag20b).insert(5, frag10b)
       # Width of spaces: 2.5 * 2 + 5 = 10  (from AFM file, adjusted for font size)
       # Line width: 20 * 4 + width_of_spaces = 90
       # Missing width: 100 - 90 = 10
       # -> Each space must be doubled!
 
-      box = HexaPDF::Layout::TextBox.new(items: items, width: 100)
-      box.style.align = :justify
-      rest, _height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: items, width: 100)
+      layouter.style.align = :justify
+      rest, _height = layouter.fit
       assert(rest.empty?)
-      assert_equal(9, box.lines[0].items.count)
-      assert_in_delta(100, box.lines[0].width)
-      assert_equal(-250, box.lines[0].items[1].items[0])
-      assert_equal(-250, box.lines[0].items[4].items[0])
-      assert_equal(-250, box.lines[0].items[6].items[0])
-      assert_equal(30, box.lines[1].width)
+      assert_equal(9, layouter.lines[0].items.count)
+      assert_in_delta(100, layouter.lines[0].width)
+      assert_equal(-250, layouter.lines[0].items[1].items[0])
+      assert_equal(-250, layouter.lines[0].items[4].items[0])
+      assert_equal(-250, layouter.lines[0].items[6].items[0])
+      assert_equal(30, layouter.lines[1].width)
     end
 
     it "applies the optional horizontal offsets if set" do
       x_offsets = lambda {|height, line_height| height + line_height}
-      box = HexaPDF::Layout::TextBox.new(items: boxes(*([[20, 10]] * 7)), width: 60,
-                                         x_offsets: x_offsets, height: 100, style: @style)
-      rest, height = box.fit
+      layouter = HexaPDF::Layout::TextLayouter.new(items: boxes(*([[20, 10]] * 7)), width: 60,
+                                                   x_offsets: x_offsets, height: 100, style: @style)
+      rest, height = layouter.fit
       assert(rest.empty?)
       assert_equal(30, height)
-      assert_equal(10, box.lines[0].x_offset)
-      assert_equal(20, box.lines[1].x_offset)
-      assert_equal(30, box.lines[2].x_offset)
+      assert_equal(10, layouter.lines[0].x_offset)
+      assert_equal(20, layouter.lines[1].x_offset)
+      assert_equal(30, layouter.lines[2].x_offset)
     end
   end
 
@@ -547,7 +548,7 @@ describe HexaPDF::Layout::TextBox do
       @frag = HexaPDF::Layout::TextFragment.create("This is some more text.\n" \
                                                    "This is some more text.", font: @font)
       @width = HexaPDF::Layout::TextFragment.create("This is some   ", font: @font).width
-      @box =  HexaPDF::Layout::TextBox.new(items: [@frag], width: @width)
+      @layouter = HexaPDF::Layout::TextLayouter.new(items: [@frag], width: @width)
       @canvas = @doc.pages.add.canvas
 
       @line1w = HexaPDF::Layout::TextFragment.create("This is some", font: @font).width
@@ -556,8 +557,8 @@ describe HexaPDF::Layout::TextBox do
 
     it "can horizontally align the contents to the left" do
       top = 100
-      @box.style.align = :left
-      @box.draw(@canvas, 5, top)
+      @layouter.style.align = :left
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5, top - @frag.y_max],
                         [5, top - @frag.y_max - @frag.height],
@@ -567,8 +568,8 @@ describe HexaPDF::Layout::TextBox do
 
     it "can horizontally align the contents to the center" do
       top = 100
-      @box.style.align = :center
-      @box.draw(@canvas, 5, top)
+      @layouter.style.align = :center
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5 + (@width - @line1w) / 2, top - @frag.y_max],
                         [5 + (@width - @line2w) / 2, top - @frag.y_max - @frag.height],
@@ -578,8 +579,8 @@ describe HexaPDF::Layout::TextBox do
 
     it "can horizontally align the contents to the right" do
       top = 100
-      @box.style.align = :right
-      @box.draw(@canvas, 5, top)
+      @layouter.style.align = :right
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5 + @width - @line1w, top - @frag.y_max],
                         [5 + @width - @line2w, top - @frag.y_max - @frag.height],
@@ -589,32 +590,32 @@ describe HexaPDF::Layout::TextBox do
 
     it "can justify the contents" do
       top = 100
-      @box.style.align = :justify
-      @box.draw(@canvas, 5, top)
+      @layouter.style.align = :justify
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5, top - @frag.y_max],
                         [5, top - @frag.y_max - @frag.height],
                         [5, top - @frag.y_max - @frag.height * 2],
                         [5, top - @frag.y_max - @frag.height * 3]])
-      assert_in_delta(@width, @box.lines[0].width, 0.0001)
-      assert_in_delta(@width, @box.lines[2].width, 0.0001)
+      assert_in_delta(@width, @layouter.lines[0].width, 0.0001)
+      assert_in_delta(@width, @layouter.lines[2].width, 0.0001)
     end
 
     it "doesn't justify lines ending in a mandatory break or the last line" do
-      @box.style.align = :justify
-      @box.draw(@canvas, 5, 100)
-      assert_equal(@line2w, @box.lines[1].width, 0.0001)
-      assert_equal(@line2w, @box.lines[3].width, 0.0001)
+      @layouter.style.align = :justify
+      @layouter.draw(@canvas, 5, 100)
+      assert_equal(@line2w, @layouter.lines[1].width, 0.0001)
+      assert_equal(@line2w, @layouter.lines[3].width, 0.0001)
     end
 
     it "can vertically align the contents in the center" do
       top = 100
-      @box = HexaPDF::Layout::TextBox.new(items: [@frag], width: @width, height: top)
-      @box.style.valign = :center
+      @layouter = HexaPDF::Layout::TextLayouter.new(items: [@frag], width: @width, height: top)
+      @layouter.style.valign = :center
 
-      _, height = @box.fit
+      _, height = @layouter.fit
       initial_baseline = top - ((top - height) / 2) - @frag.y_max
-      @box.draw(@canvas, 5, top)
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5, initial_baseline],
                         [5, initial_baseline - @frag.height],
@@ -624,12 +625,12 @@ describe HexaPDF::Layout::TextBox do
 
     it "can vertically align the contents to the bottom" do
       top = 100
-      @box = HexaPDF::Layout::TextBox.new(items: [@frag], width: @width, height: top)
-      @box.style.valign = :bottom
+      @layouter = HexaPDF::Layout::TextLayouter.new(items: [@frag], width: @width, height: top)
+      @layouter.style.valign = :bottom
 
-      _, height = @box.fit
+      _, height = @layouter.fit
       initial_baseline = height - @frag.y_max
-      @box.draw(@canvas, 5, top)
+      @layouter.draw(@canvas, 5, top)
       assert_positions(@canvas.contents,
                        [[5, initial_baseline],
                         [5, initial_baseline - @frag.height],
@@ -638,29 +639,29 @@ describe HexaPDF::Layout::TextBox do
     end
 
     it "raises an error if vertical alignment is :center/:bottom and an unlimited height is used" do
-      @box = HexaPDF::Layout::TextBox.new(items: [@frag], width: @width)
+      @layouter = HexaPDF::Layout::TextLayouter.new(items: [@frag], width: @width)
       assert_raises(HexaPDF::Error) do
-        @box.style.valign = :center
-        @box.draw(@canvas, 0, 0)
+        @layouter.style.valign = :center
+        @layouter.draw(@canvas, 0, 0)
       end
       assert_raises(HexaPDF::Error) do
-        @box.style.valign = :bottom
-        @box.draw(@canvas, 0, 0)
+        @layouter.style.valign = :bottom
+        @layouter.draw(@canvas, 0, 0)
       end
     end
 
     it "makes sure that text fragments don't pollute the graphics state for inline boxes" do
       frag = HexaPDF::Layout::TextFragment.create("Demo", font: @font)
       inline_box = HexaPDF::Layout::InlineBox.new(10, 10) {|_, canvas| canvas.text("A")}
-      box = HexaPDF::Layout::TextBox.new(items: [frag, inline_box], width: 200)
-      assert_raises(HexaPDF::Error) { box.draw(@canvas, 0, 0) }
+      layouter = HexaPDF::Layout::TextLayouter.new(items: [frag, inline_box], width: 200)
+      assert_raises(HexaPDF::Error) { layouter.draw(@canvas, 0, 0) }
     end
 
     it "doesn't do unnecessary work for placeholder boxes" do
       box1 = HexaPDF::Layout::InlineBox.new(10, 20)
       box2 = HexaPDF::Layout::InlineBox.new(30, 40) { @canvas.line_width(2) }
-      box = HexaPDF::Layout::TextBox.new(items: [box1, box2], width: 200)
-      box.draw(@canvas, 0, 0)
+      layouter = HexaPDF::Layout::TextLayouter.new(items: [box1, box2], width: 200)
+      layouter.draw(@canvas, 0, 0)
       assert_operators(@canvas.contents, [[:save_graphics_state],
                                           [:restore_graphics_state],
                                           [:save_graphics_state],